speech-service/requirements.txt
# Google Speech-to-Text Service Dependencies
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
flask==2.3.3
flask-cors==4.0.0
websockets==11.0.3
pyaudio==0.2.11
numpy==1.24.3
python-dotenv==1.0.0
